issue #24 new classes
