+++ src/main/java/com/corporate/hellscape/events/Gift.java
@@ -1,0 +1,17 @@
+package com.corporate.hellscape.events;
+import com.corporate.hellscape.Hellscape;
+
+public class Gift extends Event {
+	private boolean _triggerEnabled;
+    public Gift(boolean triggerEnabled) {
+        _triggerEnabled = _triggerEnabled; }
+    public boolean isTriggered(Hellscape hellscape) {
+        return _triggerEnabled;}
+    public void triggerAction(Hellscape hellscape) {
+        System.out.println("your friend got you a nice gift");
+		Character.Stress =- 5;
+		Character.Fun =+ 5;
+
+        hellscape.registerEvent(new Gift(false));
+    }
+}
